Switch over to use Laravel's built-in controller layouts
We were trying to reinvent the wheel when it wasn't necessary. This also
addresses errors that were thrown about sending after headers had already
been sent (which, of course, was fixed in the L4 core as I was working on
this). Hopefully this is a much cleaner way of handling the controller
work now.