Update the base model
We were trying to reinvent the wheel a bunch in here, so this is much
simpler than before. Instead of `add()`, we now use `create()` and make
sure we scrub the input data. Instead of `edit()`, we now use `update()`
and make sure we scrub the input data. Instead of `remove()`, we now use
a combination of `delete()` and `destroy()` based on what we're trying
to do.