Create a symlink from nova/vendor to /vendor if none exists
This will make sure that vendor can live in nova/vendor but
anything that by default looks for vendor folder in root will
find it still, and work as expected.

PHP's symlink works on both linux and Windows machines, so this
should be consistent, and the symlink generation will really only
happen once, unless the vendor symlink was deleted, so it is
ensured to always exist and be referenced.

Fixes #157